requirements.txt
wheel
accelerate==1.6.0
diffusers==0.33.1
googletrans==3.1.0a0
moviepy==1.0.3
#openai-whisper==20240930
#opencv-python==4.11.0.86
pillow==11.2.1
pydub==0.25.1
pysrt==1.1.2
#python-ffmpeg==2.0.12
torch==2.7.0
transformers==4.51.3
TTS==0.22.0
whisper==1.1.10
